refacto: move the bottom bar to the scaffold subpackage
