+++ app/src/main/java/com/illiouchine/jm/ui/composable/scaffold/MjuScaffold.kt
@@ -27,8 +27,6 @@ import androidx.compose.ui.tooling.preview.AndroidUiModes.UI_MODE_NIGHT_YES
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
-import com.illiouchine.jm.ui.composable.MjuBottomBar
-import com.illiouchine.jm.ui.composable.MjuNavigationRail
 import com.illiouchine.jm.ui.navigator.Screens
 import com.illiouchine.jm.ui.theme.JmTheme
 import com.illiouchine.jm.ui.theme.Theme
